refactor(#51): 음성채팅 로직 분리 및 개선
