+++ src/editor/editor.gateway.ts
@@ -40,9 +40,8 @@ export class EditorGateway implements OnGatewayConnection, OnGatewayDisconnect {
   >(); // collaborationId -> { teacherSocketId, studentSocketId }
   private collaborationSVGs = new Map<string, SVGLine[]>(); // collaborationId -> lines
 
-  // 음성채팅 관련 상태 관리
-  private voiceUsers = new Map<string, ConnectedUser>(); // socketId -> user (음성채팅 참가자)
-  private roomVoiceUsers = new Map<string, Set<string>>(); // roomId -> Set<socketId> (방별 음성채팅 참가자)
+  // (voiceUsers, roomVoiceUsers 상태 삭제)
+  // (handleVoiceJoin, handleVoiceSignal, handleVoiceLeave 메서드 전체 삭제)
 
   @WebSocketServer()
   server: Server;
@@ -66,23 +65,6 @@ export class EditorGateway implements OnGatewayConnection, OnGatewayDisconnect {
         if (users.size === 0) this.roomUsers.delete(user.roomId);
       }
       this.connectedUsers.delete(client.id);
-
-      // 음성채팅에서도 제거
-      const voiceUser = this.voiceUsers.get(client.id);
-      if (voiceUser) {
-        const voiceUsers = this.roomVoiceUsers.get(voiceUser.roomId);
-        if (voiceUsers) {
-          voiceUsers.delete(client.id);
-          if (voiceUsers.size === 0)
-            this.roomVoiceUsers.delete(voiceUser.roomId);
-        }
-        this.voiceUsers.delete(client.id);
-
-        // 방의 다른 음성채팅 참가자들에게 퇴장 알림
-        this.server.to(`room_${voiceUser.inviteCode}`).emit('voice:leave', {
-          userId: voiceUser.userId,
-        });
-      }
 
       // 해당 사용자가 참여한 협업 세션들 정리
       for (const [
@@ -172,26 +154,7 @@ export class EditorGateway implements OnGatewayConnection, OnGatewayDisconnect {
     }
     client.leave(`room_${payload.inviteCode}`);
 
-    // 2. 음성채팅에서도 제거
-    const voiceUser = this.voiceUsers.get(client.id);
-    if (voiceUser) {
-      const voiceUsers = this.roomVoiceUsers.get(voiceUser.roomId);
-      if (voiceUsers) {
-        voiceUsers.delete(client.id);
-        if (voiceUsers.size === 0) {
-          this.roomVoiceUsers.delete(voiceUser.roomId);
-        }
-      }
-      this.voiceUsers.delete(client.id);
-
-      // 방의 다른 음성채팅 참가자들에게 퇴장 알림
-      this.server.to(`room_${payload.inviteCode}`).emit('voice:leave', {
-        userId: voiceUser.userId,
-      });
-      console.log(`음성채팅 퇴장: ${voiceUser.userName} (${voiceUser.userId})`);
-    }
-
-    // 3. DB에서 participants에서 해당 학생 제거
+    // 2. DB에서 participants에서 해당 학생 제거
     const room = await this.roomRepository.findOne({
       where: { inviteCode: payload.inviteCode },
     });
@@ -206,7 +169,7 @@ export class EditorGateway implements OnGatewayConnection, OnGatewayDisconnect {
       await this.roomRepository.save(room);
     }
 
-    // 4. 방 전체에 room:updated broadcast
+    // 3. 방 전체에 room:updated broadcast
     void this.server.to(`room_${payload.inviteCode}`).emit('room:updated');
     console.log('room:updated emit (leave)', payload.inviteCode);
   }
@@ -404,144 +367,5 @@ export class EditorGateway implements OnGatewayConnection, OnGatewayDisconnect {
   }
 
   // 음성채팅 관련 이벤트 핸들러들
-  @SubscribeMessage('voice:join')
-  handleVoiceJoin(
-    @ConnectedSocket() client: Socket,
-    @MessageBody()
-    payload: {
-      roomId: string;
-      userId: number;
-      userName: string;
-      userRole: 'teacher' | 'student';
-    },
-  ) {
-    console.log('voice:join 수신', payload, client.id);
-
-    // 사용자 정보 가져오기
-    const user = this.connectedUsers.get(client.id);
-    if (!user) {
-      console.log('음성채팅 참가 실패: 사용자가 방에 입장하지 않음', client.id);
-      return;
-    }
-
-    // 기존 참가자 목록을 먼저 생성 (자신을 추가하기 전에)
-    let voiceUsers = this.roomVoiceUsers.get(payload.roomId);
-    if (!voiceUsers) {
-      voiceUsers = new Set();
-      this.roomVoiceUsers.set(payload.roomId, voiceUsers);
-    }
-
-    const existingParticipants = Array.from(voiceUsers)
-      .map((socketId) => {
-        const participant = this.voiceUsers.get(socketId);
-        return participant
-          ? {
-              userId: participant.userId,
-              userName: participant.userName,
-              userRole: participant.role,
-            }
-          : null;
-      })
-      .filter(Boolean);
-
-    // 기존 참가자 목록 전송
-    if (existingParticipants.length > 0) {
-      this.server
-        .to(client.id)
-        .emit('voice:existing-participants', existingParticipants);
-      console.log('기존 참가자 목록 전송:', existingParticipants);
-    }
-
-    // 이제 자신을 음성채팅 참가자로 등록
-    this.voiceUsers.set(client.id, user);
-    voiceUsers.add(client.id);
-
-    // 방의 다른 음성채팅 참가자들에게 새 참가자 알림
-    this.server.to(`room_${user.inviteCode}`).emit('voice:user-joined', {
-      userId: user.userId,
-      userName: user.userName,
-      userRole: user.role,
-    });
-
-    console.log(
-      `${user.role} ${user.userName} joined voice chat in room ${user.inviteCode}`,
-    );
-  }
-
-  @SubscribeMessage('voice:signal')
-  handleVoiceSignal(
-    @ConnectedSocket() client: Socket,
-    @MessageBody()
-    payload: {
-      roomId: string;
-      to: number; // 대상 사용자 ID
-      signal: Record<string, unknown>; // WebRTC 시그널링 데이터
-    },
-  ) {
-    console.log('voice:signal 수신', { from: client.id, to: payload.to });
-
-    // 대상 사용자의 socketId 찾기
-    const targetUser = [...this.voiceUsers.entries()].find(
-      ([, user]) =>
-        user.userId === payload.to && user.roomId === payload.roomId,
-    );
-
-    if (targetUser) {
-      const [targetSocketId] = targetUser;
-
-      // 시그널링 데이터를 대상 사용자에게 전달
-      this.server.to(targetSocketId).emit('voice:signal', {
-        from: this.voiceUsers.get(client.id)?.userId,
-        signal: payload.signal,
-      });
-
-      console.log(
-        `음성 시그널링: ${client.id} -> ${targetSocketId} (${payload.to})`,
-      );
-    } else {
-      console.log(`음성채팅 대상 사용자를 찾을 수 없음: userId=${payload.to}`);
-    }
-  }
-
-  @SubscribeMessage('voice:leave')
-  handleVoiceLeave(
-    @ConnectedSocket() client: Socket,
-    @MessageBody()
-    payload: {
-      roomId: string;
-      userId: number;
-    },
-  ) {
-    console.log('voice:leave 수신', payload, client.id);
-
-    // 사용자 정보 가져오기
-    const user = this.voiceUsers.get(client.id);
-    if (!user) {
-      console.log(
-        '음성채팅 퇴장 실패: 사용자가 음성채팅에 참가하지 않음',
-        client.id,
-      );
-      return;
-    }
-
-    // 음성채팅 참가자에서 제거
-    this.voiceUsers.delete(client.id);
-
-    const voiceUsers = this.roomVoiceUsers.get(payload.roomId);
-    if (voiceUsers) {
-      voiceUsers.delete(client.id);
-      if (voiceUsers.size === 0) {
-        this.roomVoiceUsers.delete(payload.roomId);
-      }
-    }
-
-    // 방의 다른 음성채팅 참가자들에게 퇴장 알림
-    this.server.to(`room_${user.inviteCode}`).emit('voice:leave', {
-      userId: user.userId,
-    });
-
-    console.log(
-      `${user.role} ${user.userName} left voice chat in room ${user.inviteCode}`,
-    );
-  }
+  // (handleVoiceJoin, handleVoiceSignal, handleVoiceLeave 메서드 전체 삭제)
 }
